clone-your-own-voice/requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
elevenlabs==1.54.0
google-generativeai==0.3.2
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27
audio-recorder-streamlit==0.0.8
numpy==1.26.4 